fix(small): fix scroll in firefox, fix after pixelperfext review, add useScroll hook
