+++ src/components/CandidatesTable/CandidatesTable.jsx
@@ -1,10 +1,10 @@
 import './CandidatesTable.scss';
 
-import { useRef, useEffect, useState, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 
 import PropTypes from 'prop-types';
 
-import { debounce } from '../../utils/utils';
+import { useScroll } from '../../utils/hooks/useScroll';
 
 import CandidatesTableCell from '../CandidatesTableCell/CandidatesTableCell';
 import CandidatesTableSort from '../CandidatesTableSort/CandidatesTableSort';
@@ -12,25 +12,10 @@ import CandidatesTableSort from '../CandidatesTableSort/CandidatesTableSort';
 const CandidatesTable = (props) => {
   const { candidates } = props;
 
-  const [scrollPosition, setScrollPosition] = useState(0);
-  const ref = useRef(null);
   const [sortCandidates, setSortCandidates] = useState(candidates);
   const [candidatesList, setCandidatesList] = useState(null);
 
-  const handleSCrollPosition = useCallback(() => {
-    setScrollPosition(document.querySelector('#candidates-table').scrollTop);
-  }, []);
-
-  // Чтобы функция вызывалась не чаще чем раз в 300 мсек
-  const setScrollPositionDebaunced = debounce(handleSCrollPosition, 300);
-
-  // Слушатель события на появление скролла у контейнера с кандидатами
-  useEffect(() => {
-    document
-      .querySelector('#candidates-table')
-      .addEventListener('scroll', setScrollPositionDebaunced);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const { scrollToTop, scrollPosition, ref } = useScroll(`#candidates-table`);
 
   useEffect(() => {
     setSortCandidates(candidates);
@@ -46,15 +31,7 @@ const CandidatesTable = (props) => {
     );
   }, [sortCandidates]);
 
-  // Передвигаемся в начало контейнера с кандидатами
-  const scrollToTop = () => {
-    const firstChildElement = ref.current?.firstElementChild;
-    firstChildElement?.scrollIntoView({ behavior: 'smooth' });
-  };
-
   const handleSortCandidates = (sortValue) => {
-    // console.log(sortValue);
-
     setSortCandidates((prevState) => {
       const sortedCandidates = [...prevState];
 
@@ -87,7 +64,6 @@ const CandidatesTable = (props) => {
       } else if (sortValue.work === 'sortDown') {
         sortedCandidates.sort((a, b) => a.work.localeCompare(b.work));
       }
-      // console.log('click', sortCandidates);
       return sortedCandidates;
     });
   };
