+++ src/components/UI/MatchStatus/MatchStatus.scss
@@ -1,0 +1,29 @@
+@import '../../Shared/mixins.scss';
+@import '../../Shared/variables.scss';
+
+.match-status {
+  @include font-sans($textXS, $color-success, 300, 120%, 0.24px);
+  padding-left: 22px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: $color-bg-green;
+  background-image: url(../../../images/icons/icon-vacancies-table/arrow_up_green.svg);
+  background-repeat: no-repeat;
+  background-position: left 11px top 6px;
+  border-radius: 6px;
+  width: 52px;
+  height: 30px;
+
+  &_color_blue {
+    @include font-sans($textXS, $color-blue, 300, 120%, 0.24px);
+    background-color: $color-bg-blue;
+    background-image: url(../../../images/icons/icon-vacancies-table/arrow_up_blue.svg);
+  }
+
+  &_color_red {
+    @include font-sans($textXS, $color-error, 300, 120%, 0.24px);
+    background-color: $color-bg-red;
+    background-image: url(../../../images/icons/icon-vacancies-table/arrow_down_red.svg);
+  }
+}
