+++ src/components/VacanciesPage/VacanciesPage.scss
@@ -3,6 +3,7 @@
 
 .vacanсies-page {
   width: 1025px;
+  padding-right: 32px;
 
   &__header {
     width: 1025px;
